+++ app/index.js
@@ -6,7 +6,6 @@ import { Provider } from 'react-redux';
 import { useScreens } from 'react-native-screens'; // eslint-disable-line import/no-unresolved
 import { Linking } from 'react-native';
 import PropTypes from 'prop-types';
-import { Transition, Transitioning } from 'react-native-reanimated';
 
 import { appInit } from './actions';
 import { deepLinkingOpen } from './actions/deepLinking';
@@ -20,9 +19,7 @@ import { defaultHeader, onNavigationStateChange } from './utils/navigation';
 import { loggerConfig, analytics } from './utils/log';
 import Toast from './containers/Toast';
 import RocketChat from './lib/rocketchat';
-import debounce from './utils/debounce';
-import sharedStyles from './views/Styles';
-import { isIOS, isAndroid } from './utils/deviceInfo';
+import LayoutAnimation from './utils/layoutAnimation';
 
 useScreens();
 
@@ -264,22 +261,6 @@ const App = createAppContainer(createSwitchNavigator(
 	}
 ));
 
-const transition = (
-	<Transition.Together>
-		<Transition.In type='fade' />
-		<Transition.Out type='fade' />
-		<Transition.Change interpolation='easeInOut' />
-	</Transition.Together>
-);
-
-const TRANSITION_REF = React.createRef();
-
-export const animateNextTransition = debounce(() => {
-	if (isIOS) {
-		TRANSITION_REF.current.animateNextTransition();
-	}
-}, 200, true);
-
 export default class Root extends React.Component {
 	constructor(props) {
 		super(props);
@@ -326,32 +307,16 @@ export default class Root extends React.Component {
 	}
 
 	render() {
-		if (isAndroid) {
-			return (
-				<Provider store={store}>
-					<App
-						ref={(navigatorRef) => {
-							Navigation.setTopLevelNavigator(navigatorRef);
-						}}
-						onNavigationStateChange={onNavigationStateChange}
-					/>
-				</Provider>
-			);
-		}
 		return (
 			<Provider store={store}>
-				<Transitioning.View
-					style={sharedStyles.root}
-					transition={transition}
-					ref={TRANSITION_REF}
-				>
+				<LayoutAnimation>
 					<App
 						ref={(navigatorRef) => {
 							Navigation.setTopLevelNavigator(navigatorRef);
 						}}
 						onNavigationStateChange={onNavigationStateChange}
 					/>
-				</Transitioning.View>
+				</LayoutAnimation>
 			</Provider>
 		);
 	}
